+++ src/Snap/Internal/Http/Server/SimpleBackend.hs
@@ -91,7 +91,7 @@ name :: ByteString
 name = "simple"
 
 
-sendFile :: Connection -> FilePath -> Int -> IO ()
+sendFile :: Connection -> FilePath -> Int64 -> IO ()
 #if defined(HAS_SENDFILE)
 sendFile c fp sz = do
     fd <- openFd fp ReadOnly Nothing defaultFileFlags
